codex_artifacts/kits/test-3.18/requirements.txt
e2b==2.15.2
e2b-code-interpreter
python-dotenv
